Move test to compat for v11
