+++ test/render.test.jsx
@@ -394,27 +394,6 @@ describe('render', () => {
 				match({})
 			);
 		});
-
-		it('should apply defaultProps', () => {
-			const Test = (props) => <div {...props} />;
-			Test.defaultProps = {
-				foo: 'default foo',
-				bar: 'default bar'
-			};
-
-			expect(render(<Test />), 'defaults').to.equal(
-				'<div foo="default foo" bar="default bar"></div>'
-			);
-			expect(render(<Test bar="b" />), 'partial').to.equal(
-				'<div bar="b" foo="default foo"></div>'
-			);
-			expect(render(<Test foo="a" bar="b" />), 'overridden').to.equal(
-				'<div foo="a" bar="b"></div>'
-			);
-			expect(render(<Test foo={undefined} bar="b" />), 'overridden').to.equal(
-				'<div foo="default foo" bar="b"></div>'
-			);
-		});
 	});
 
 	describe('Classical Components', () => {
@@ -485,31 +464,6 @@ describe('render', () => {
 				}),
 				match({}),
 				match({})
-			);
-		});
-
-		it('should apply defaultProps', () => {
-			class Test extends Component {
-				render(props) {
-					return <div {...props} />;
-				}
-			}
-			Test.defaultProps = {
-				foo: 'default foo',
-				bar: 'default bar'
-			};
-
-			expect(render(<Test />), 'defaults').to.equal(
-				'<div foo="default foo" bar="default bar"></div>'
-			);
-			expect(render(<Test bar="b" />), 'partial').to.equal(
-				'<div bar="b" foo="default foo"></div>'
-			);
-			expect(render(<Test foo="a" bar="b" />), 'overridden').to.equal(
-				'<div foo="a" bar="b"></div>'
-			);
-			expect(render(<Test foo={undefined} bar="b" />), 'overridden').to.equal(
-				'<div foo="default foo" bar="b"></div>'
 			);
 		});
